Version 2.6.20 - Bug Fixed - Parameter type changed

+++ src/main/java/com/kalsym/deliveryservice/models/daos/DeliveryRemarks.java
@@ -25,5 +25,5 @@ public class DeliveryRemarks implements Serializable {
     String title;
     String message;
     String deliveryType;
-    Long providerId;
+    Integer providerId;
 }
